SSDM-3539 : Bugfix, snapshots should create a package.
SVN: 36239

+++ openbis_all/source/bash/create-snapshot.sh
@@ -101,6 +101,11 @@ for db in $DATABASES; do
     fi
     echo "Database '$db' has been successfully dumped."
 done
+############## packaging ##############
+if ! tar -zcf "$SNAPSHOT.tgz" -C "$REPOSITORY" "$SNAPSHOT_FOLDER_NAME"; then
+    echo "Error packaging snapshot $SNAPSHOT."
+    exit 1
+fi
 rm -rf "$SNAPSHOT"
 
 echo "==== $SNAPSHOT.tgz successfully created."
